fix tests and move call to oapi to get recipients to routes, remove call to ordapi for selected recipients
