+++ app/pages/sections/order-items/catalogue-solutions/select/recipients/manifest.json
@@ -1,8 +1,7 @@
 {
   "backLinkText": "Go back",
-  "title": "Service Recipients for",
+  "title": "Service Recipients for {{itemName}} for {{orderId}}",
   "description": "These are all the possible Service Recipients for this Call-off Ordering Party. Select those you’ll be ordering Catalogue Solutions and Additional Services for.",
-  "insetAdvice": "You only need to select Service Recipients if you intend to order any Catalogue Solutions or Additional Services as part of this order. If you do not want to add any, or have finished adding Service Recipients, continue to the next step.",
   "selectDeselectButtonText": {
     "select": "Select all",
     "deselect": "Deselect all"
