+++ app/pages/sections/order-items/catalogue-solutions/select/recipients/template.njk
@@ -14,15 +14,10 @@
       }) }}
     </div>
     <div class="nhsuk-grid-row">
+    
       <div class="nhsuk-grid-column-two-thirds">
         <h1 data-test-id="solution-recipients-page-title" class="nhsuk-u-margin-top-7 nhsuk-u-margin-bottom-3">{{ title }}</h1>
         <h2 data-test-id="solution-recipients-page-description" class="nhsuk-u-font-size-24 nhsuk-u-margin-bottom-7">{{ description }}</h2>
-
-        <div data-test-id="solution-recipients-page-insetAdvice">
-          {{ insetText({
-            HTML: insetAdvice
-          }) }}
-        </div>
       </div>
 
       <div class="nhsuk-grid-column-full">
